@@ -6,7 +6,10 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
+
 import androidx.activity.EdgeToEdge;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
@@ -20,6 +23,8 @@ import java.security.NoSuchProviderException;
 import java.security.UnrecoverableEntryException;
 import java.security.cert.CertificateException;
 import java.security.spec.InvalidKeySpecException;
+import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
 
 import com.scottyab.rootbeer.RootBeer;
 
@@ -27,6 +32,19 @@ import com.scottyab.rootbeer.RootBeer;
 public class MainActivity extends AppCompatActivity {
 
     private KeyManager keyManager;
+
+    private ActivityResultLauncher<Intent> cameraActivityResultLauncher;
+
+    private Utilities utilities;
+
+    private int id;
+
+    private String timestamp;
+
+
+    private void prepareDataForSending(String timestamp) throws Exception {
+        utilities.beginImageEncryption(timestamp);
+    }
 
 
     private void ManageKey() throws InvalidAlgorithmParameterException, UnrecoverableEntryException, CertificateException, NoSuchAlgorithmException, KeyStoreException, IOException, InvalidKeySpecException, NoSuchProviderException {
@@ -47,8 +65,6 @@ public class MainActivity extends AppCompatActivity {
         RootBeer rootBeer = new RootBeer(this);
 
 
-
-
         if (rootBeer.isRooted()) {
             Toast.makeText(this, "Device is rooted! Exiting app...", Toast.LENGTH_LONG).show();
             finishAffinity();
@@ -56,22 +72,50 @@ public class MainActivity extends AppCompatActivity {
 
         keyManager = new KeyManager(this);
 
+        utilities = new Utilities(this, id);
+
+        Intent intent = getIntent();
+        if (Objects.equals(intent.getAction(), "com.auth.ACTION_REQUEST_DATA")) {
+            id = intent.getIntExtra("ID", 0);
+        }
+
         Button cameraButton = findViewById(R.id.Camera);
 
         try {
             ManageKey();
-        } catch (InvalidAlgorithmParameterException | UnrecoverableEntryException | CertificateException | NoSuchAlgorithmException | KeyStoreException | IOException | InvalidKeySpecException | NoSuchProviderException e) {
+        } catch (InvalidAlgorithmParameterException | UnrecoverableEntryException |
+                 CertificateException | NoSuchAlgorithmException | KeyStoreException | IOException |
+                 InvalidKeySpecException | NoSuchProviderException e) {
             Log.e("MainActivity", "Error managing key: ", e);
         }
 
-        cameraButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent cameraIntent = new Intent(MainActivity.this, CameraActivity.class);
-                startActivity(cameraIntent);
-            }
-        });
+        cameraActivityResultLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(),
+                result -> {
+                    if (result.getResultCode() == RESULT_OK) {
+                        Intent data = result.getData();
+                        assert data != null;
+                        timestamp = data.getStringExtra("timestamp");
 
+                        CompletableFuture.runAsync(() -> utilities.downloadPublicKey())
+                                .thenRun(() -> {
+                                    try {
+                                        prepareDataForSending(timestamp);
+                                    } catch (Exception e) {
+                                        throw new RuntimeException(e);
+                                    }
+                                });
+
+                    }
+                });
+
+        cameraButton.setOnClickListener(this::openCameraActivity);
+
+    }
+
+    public void openCameraActivity(View view) {
+        Intent intent = new Intent(this, CameraActivity.class);
+        intent.putExtra("ID", id);
+        cameraActivityResultLauncher.launch(intent);
     }
 
 }
